refactor(router): extract pure pattern logic into url module
