+++ packages/router/src/index.ts
@@ -1,5 +1,6 @@
-export { Router, matchPattern } from './router';
-export type { Match } from './router';
+export { matchPattern, fullPattern, patternSegment, specificity } from './url';
+export type { Match } from './url';
+export { Router, resolveChild } from './router';
 export { Route } from './route';
 export { Link } from './link';
 export { Redirect } from './redirect';
